webapp/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
mysql-connector-python==8.0.33
Flask-SQLAlchemy==3.0.5
psutil==5.9.8
prometheus-client==0.17.1
